refactor: fix folder names
