+++ src/app/types/sectionItem.ts
